+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/SpotlessExtension.java
@@ -91,6 +91,12 @@ public class SpotlessExtension {
 		configure(KotlinExtension.NAME, KotlinExtension.class, closure);
 	}
 
+	/** Configures the special Gradle Kotlin DSL specific extension. */
+	public void kotlinGradle(Action<KotlinGradleExtension> closure) {
+		requireNonNull(closure);
+		configure(KotlinGradleExtension.NAME, KotlinGradleExtension.class, closure);
+	}
+
 	/** Configures the special freshmark-specific extension. */
 	public void freshmark(Action<FreshMarkExtension> closure) {
 		requireNonNull(closure);